Revise probe expression for the unwrap_infallible feature.
This allows the probe (compiling of the expression) to succeed, when the
unstable feature is enabled with the current nightly rustc.  This change is
currently needed because `Infallible: Into<!>` (as required by Result::into_ok)
is not yet supported by Rust.

This also seems to follow the API of the feature better, since its design chose
to involve `!` instead of Infallible.