Improve dev testing with the ability to exercise unstable features.
The ability is disabled (i.e. commented-out) by default, because it requires
hacks that the package must not normally involve.  The `cfg` option that is part
of enabling it is intentionally not a Cargo feature.

The ability currently requires using my hacked fork of the `autocfg` crate (only
when the ability is specially enabled (i.e. uncommented)).  This takes advantage
of Cargo config files applying to everything under their directory, so that we
only need to specify `[patch.crates-io]` in one place.  This hacked fork is now
provided as a Git submodule, so that any future changes to it will be kept
in-sync.